move pause & restart button to button panel
