+++ WPFUI/ViewModels/Uc/MainView/MainButtonPanelViewModel.cs
@@ -1,26 +1,27 @@
 ﻿using MainCore.Enums;
 using MainCore.Helper;
 using MainCore.Tasks.Misc;
+using MainCore.Tasks.Sim;
 using ReactiveUI;
 using Splat;
 using System;
 using System.Linq;
 using System.Reactive;
+using System.Reactive.Concurrency;
 using System.Reactive.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 using WPFUI.Models;
 using WPFUI.ViewModels.Abstract;
-using WPFUI.ViewModels.Uc.MainView;
 using Access = MainCore.Models.Database.Access;
 
-namespace WPFUI.ViewModels.Uc
+namespace WPFUI.ViewModels.Uc.MainView
 {
-    public class ButtonPanelViewModel : AccountTabBaseViewModel
+    public class MainButtonPanelViewModel : AccountTabBaseViewModel
     {
         private readonly MainTabPanelViewModel _mainTabPanelViewModel;
 
-        public ButtonPanelViewModel()
+        public MainButtonPanelViewModel()
         {
             _eventManager.AccountStatusUpdate += OnAccountUpdate;
             _mainTabPanelViewModel = Locator.Current.GetService<MainTabPanelViewModel>();
@@ -32,8 +33,9 @@ namespace WPFUI.ViewModels.Uc
 
             LoginCommand = ReactiveCommand.CreateFromTask(LoginTask, this.WhenAnyValue(vm => vm.IsAllowLogin, vm => vm._selectorViewModel.IsAccountSelected, (a, b) => a && b));
             LogoutCommand = ReactiveCommand.CreateFromTask(LogoutTask, this.WhenAnyValue(vm => vm.IsAllowLogout, vm => vm._selectorViewModel.IsAccountSelected, (a, b) => a && b));
-            LoginAllCommand = ReactiveCommand.CreateFromTask(LoginAllTask);
-            LogoutAllCommand = ReactiveCommand.CreateFromTask(LogoutAllTask);
+
+            PauseCommand = ReactiveCommand.CreateFromTask(PauseTask, this.WhenAnyValue(x => x.IsValidStatus));
+            RestartCommand = ReactiveCommand.Create(RestartTask, this.WhenAnyValue(x => x.IsValidRestart));
         }
 
         protected override void Init(int accountId)
@@ -41,48 +43,61 @@ namespace WPFUI.ViewModels.Uc
             LoadData(accountId);
         }
 
-        private void LoadData(int accountId)
-        {
-            var status = _taskManager.GetAccountStatus(accountId);
-            switch (status)
-            {
-                case AccountStatus.Offline:
-                    IsAllowLogin = true;
-                    IsAllowLogout = false;
-                    break;
-
-                case AccountStatus.Starting:
-                    IsAllowLogin = false;
-                    IsAllowLogout = false;
-                    break;
-
-                case AccountStatus.Online:
-                    IsAllowLogin = false;
-                    IsAllowLogout = true;
-                    break;
-
-                case AccountStatus.Pausing:
-                    IsAllowLogin = false;
-                    IsAllowLogout = false;
-                    break;
-
-                case AccountStatus.Paused:
-                    IsAllowLogin = false;
-                    IsAllowLogout = true;
-                    break;
-
-                case AccountStatus.Stopping:
-                    IsAllowLogin = false;
-                    IsAllowLogout = false;
-                    break;
-            }
-        }
-
         private void OnAccountUpdate(int accountId)
         {
             if (!IsActive) return;
             if (AccountId != accountId) return;
             LoadData(accountId);
+        }
+
+        private void LoadData(int accountId)
+        {
+            var status = _taskManager.GetAccountStatus(accountId);
+
+            RxApp.MainThreadScheduler.Schedule(() =>
+            {
+                switch (status)
+                {
+                    case AccountStatus.Offline:
+                        IsAllowLogin = true;
+                        IsAllowLogout = false;
+                        break;
+
+                    case AccountStatus.Starting:
+                        IsAllowLogin = false;
+                        IsAllowLogout = false;
+                        break;
+
+                    case AccountStatus.Online:
+                        IsAllowLogin = false;
+                        IsAllowLogout = true;
+
+                        IsValidStatus = true;
+                        break;
+
+                    case AccountStatus.Pausing:
+                        IsAllowLogin = false;
+                        IsAllowLogout = false;
+                        break;
+
+                    case AccountStatus.Paused:
+                        IsAllowLogin = false;
+                        IsAllowLogout = true;
+
+                        IsValidStatus = true;
+
+                        break;
+
+                    case AccountStatus.Stopping:
+                        IsAllowLogin = false;
+                        IsAllowLogout = false;
+
+                        IsValidStatus = false;
+                        break;
+                }
+
+                IsValidRestart = status == AccountStatus.Paused;
+            });
         }
 
         private void CheckVersionTask()
@@ -104,26 +119,6 @@ namespace WPFUI.ViewModels.Uc
 
         private Task LogoutTask() => Task.Run(() => LogoutAccount(AccountId));
 
-        private async Task LoginAllTask()
-        {
-            using var context = _contextFactory.CreateDbContext();
-            var accounts = context.Accounts;
-            foreach (var account in accounts)
-            {
-                await Task.Run(() => LoginAccount(account.Id));
-            }
-        }
-
-        private async Task LogoutAllTask()
-        {
-            using var context = _contextFactory.CreateDbContext();
-            var accounts = context.Accounts;
-            foreach (var account in accounts)
-            {
-                await Task.Run(() => LogoutAccount(account.Id));
-            }
-        }
-
         private void DeleteAccountTask()
         {
             _waitingWindow.Show("saving data");
@@ -131,6 +126,10 @@ namespace WPFUI.ViewModels.Uc
             _eventManager.OnAccountsUpdate();
             _waitingWindow.Close();
         }
+
+        public Task PauseTask() => Pause(AccountId);
+
+        public void RestartTask() => Restart(AccountId);
 
         private void LoginAccount(int index)
         {
@@ -191,7 +190,7 @@ namespace WPFUI.ViewModels.Uc
             if (sleepExist is null)
             {
                 (var min, var max) = (setting.WorkTimeMin, setting.WorkTimeMax);
-                var time = TimeSpan.FromMinutes(rand.Next(min, max));
+                var time = TimeSpan.FromMinutes(Random.Shared.Next(min, max));
                 _taskManager.Add(index, new SleepTask(index) { ExecuteAt = DateTime.Now.Add(time) });
             }
 
@@ -221,14 +220,69 @@ namespace WPFUI.ViewModels.Uc
             context.SaveChanges();
         }
 
+        private async Task Pause(int index)
+        {
+            var status = _taskManager.GetAccountStatus(index);
+            if (status == AccountStatus.Paused)
+            {
+                _taskManager.UpdateAccountStatus(index, AccountStatus.Online);
+                return;
+            }
+
+            if (status == AccountStatus.Online)
+            {
+                var current = _taskManager.GetCurrentTask(index);
+                _taskManager.UpdateAccountStatus(index, AccountStatus.Pausing);
+                if (current is not null)
+                {
+                    current.Cts.Cancel();
+                    _waitingWindow.Show("waiting current task stops");
+                    await Task.Run(() =>
+                    {
+                        while (current.Stage != TaskStage.Waiting) { }
+                    });
+                    _waitingWindow.Close();
+                }
+                _taskManager.UpdateAccountStatus(index, AccountStatus.Paused);
+                return;
+            }
+        }
+
+        private void Restart(int index)
+        {
+            using var context = _contextFactory.CreateDbContext();
+            var villages = context.Villages.Where(x => x.AccountId == index);
+            _taskManager.Clear(index);
+            foreach (var village in villages)
+            {
+                var queue = _planManager.GetList(village.Id);
+                if (queue.Any())
+                {
+                    _taskManager.Add(index, new UpgradeBuilding(village.Id, index));
+                }
+                var villageSetting = context.VillagesSettings.Find(village.Id);
+                if (villageSetting.IsAutoRefresh)
+                {
+                    _taskManager.Add(index, new RefreshVillage(village.Id, index));
+                }
+            }
+
+            var setting = context.AccountsSettings.Find(index);
+            (var min, var max) = (setting.WorkTimeMin, setting.WorkTimeMax);
+
+            var time = TimeSpan.FromMinutes(Random.Shared.Next(min, max));
+            _taskManager.Add(index, new SleepTask(index) { ExecuteAt = DateTime.Now.Add(time) });
+            _taskManager.UpdateAccountStatus(index, AccountStatus.Online);
+        }
+
         public ReactiveCommand<Unit, Unit> CheckVersionCommand { get; }
         public ReactiveCommand<Unit, Unit> AddAccountCommand { get; }
         public ReactiveCommand<Unit, Unit> AddAccountsCommand { get; }
         public ReactiveCommand<Unit, Unit> DeleteAccountCommand { get; }
         public ReactiveCommand<Unit, Unit> LoginCommand { get; }
         public ReactiveCommand<Unit, Unit> LogoutCommand { get; }
-        public ReactiveCommand<Unit, Unit> LoginAllCommand { get; }
-        public ReactiveCommand<Unit, Unit> LogoutAllCommand { get; }
+        public ReactiveCommand<Unit, Unit> PauseCommand { get; }
+        public ReactiveCommand<Unit, Unit> RestartCommand { get; }
 
         private bool _isAllowLogout;
 
@@ -246,6 +300,20 @@ namespace WPFUI.ViewModels.Uc
             set => this.RaiseAndSetIfChanged(ref _isAllowLogin, value);
         }
 
-        private readonly Random rand = new();
+        private bool _isValidStatus;
+
+        public bool IsValidStatus
+        {
+            get => _isValidStatus;
+            set => this.RaiseAndSetIfChanged(ref _isValidStatus, value);
+        }
+
+        private bool _isValidRestart;
+
+        public bool IsValidRestart
+        {
+            get => _isValidRestart;
+            set => this.RaiseAndSetIfChanged(ref _isValidRestart, value);
+        }
     }
 }
