--- WPFUI/Views/Tabs/GeneralPage.xaml
@@ -1,38 +1,0 @@
-﻿<reactiveui:ReactivePage
-      x:Class="WPFUI.Views.Tabs.GeneralPage"
-      x:TypeArguments="local:GeneralViewModel"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:local="clr-namespace:WPFUI.ViewModels.Tabs"
-      xmlns:reactiveui="http://reactiveui.net"
-      mc:Ignorable="d"
-      d:DesignHeight="450" d:DesignWidth="800"
-      Title="GeneralPage">
-
-    <Grid Background="#FFE5E5E5" Margin="0,2,0,0">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="1*"/>
-            <RowDefinition Height="2*"/>
-            <RowDefinition Height="1*"/>
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="1*"/>
-            <ColumnDefinition Width="2*"/>
-            <ColumnDefinition Width="1*"/>
-        </Grid.ColumnDefinitions>
-        <Grid Grid.Column="1" Grid.Row="1" Margin="10,10,10,10">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="3*"/>
-                <RowDefinition Height="3*"/>
-                <RowDefinition Height="1*"/>
-                <RowDefinition Height="1*"/>
-            </Grid.RowDefinitions>
-            <Button x:Name="PauseButton" Content="Pause" Margin="0,0,0,10"/>
-            <Button x:Name="RestartButton" Content="Restart" Grid.Row="1" Margin="0,0,0,10"/>
-            <TextBlock HorizontalAlignment="Center" Grid.Row="2" TextWrapping="Wrap" Text="Status" VerticalAlignment="Center" FontWeight="Bold"/>
-            <TextBlock x:Name="StatusText" HorizontalAlignment="Center" Grid.Row="3" TextWrapping="Wrap" VerticalAlignment="Center" Text="Offline"/>
-        </Grid>
-    </Grid>
-</reactiveui:ReactivePage>
